Introduced the Slugger Object Service
Refactored the object graph to split logic that generates the slug from
the Listener into an object service named the Slugger. This allows other
developers to implement their own slug generation and specify using the
Service XML (DI). This allows authors the flexibility to create slugs
based on the known fields (e.g. in example; "example-document-by-cam")
or handle collisions differently (e.g. in example;
"example-document-2-by-cam"). This also allows for better independent
testing of the logic around generating of slugs, as you can simply
create the service, punch in test data (e.g. bad characters etc) and
test the result.  Updated the documentation, introduced a
SluggerInterface and implementation of Slugger (based on the slugify
doctrine method). Removed the static implementation of the collection of
collisions and passed through to the slugger.